CLI: Factor out offline helpers (bp #12382)
