+++ cli-output/Cargo.toml
@@ -1,7 +1,7 @@
 [package]
-authors = ["Solana Maintainers <maintainers@solana.com>"]
+authors = ["Solana Maintainers <maintainers@solana.foundation>"]
 edition = "2018"
-name = "solana-watchtower"
+name = "solana-cli-output"
 description = "Blockchain, Rebuilt for Scale"
 version = "1.2.30"
 repository = "https://github.com/solana-labs/solana"
@@ -9,24 +9,20 @@ license = "Apache-2.0"
 homepage = "https://solana.com/"
 
 [dependencies]
-clap = "2.33.1"
-log = "0.4.8"
+chrono = { version = "0.4.11", features = ["serde"] }
+console = "0.10.1"
 humantime = "2.0.0"
+Inflector = "0.11.4"
+indicatif = "0.14.0"
+serde = "1.0.110"
+serde_derive = "1.0.103"
+serde_json = "1.0.53"
 solana-clap-utils = { path = "../clap-utils", version = "1.2.30" }
-solana-cli-config = { path = "../cli-config", version = "1.2.30" }
-solana-cli-output = { path = "../cli-output", version = "1.2.30" }
 solana-client = { path = "../client", version = "1.2.30" }
-solana-logger = { path = "../logger", version = "1.2.30" }
-solana-metrics = { path = "../metrics", version = "1.2.30" }
-solana-notifier = { path = "../notifier", version = "1.2.30" }
 solana-sdk = { path = "../sdk", version = "1.2.30" }
+solana-stake-program = { path = "../programs/stake", version = "1.2.30" }
 solana-transaction-status = { path = "../transaction-status", version = "1.2.30" }
-solana-version = { path = "../version", version = "1.2.30" }
 solana-vote-program = { path = "../programs/vote", version = "1.2.30" }
-
-[[bin]]
-name = "solana-watchtower"
-path = "src/main.rs"
 
 [package.metadata.docs.rs]
 targets = ["x86_64-unknown-linux-gnu"]
