+++ cli-output/src/lib.rs
@@ -1,0 +1,3 @@
+mod cli_output;
+pub mod display;
+pub use cli_output::*;
