+++ cli/Cargo.toml
@@ -31,6 +31,7 @@ solana-account-decoder = { path = "../account-decoder", version = "1.2.30" }
 solana-budget-program = { path = "../programs/budget", version = "1.2.30" }
 solana-clap-utils = { path = "../clap-utils", version = "1.2.30" }
 solana-cli-config = { path = "../cli-config", version = "1.2.30" }
+solana-cli-output = { path = "../cli-output", version = "1.2.30" }
 solana-client = { path = "../client", version = "1.2.30" }
 solana-config-program = { path = "../programs/config", version = "1.2.30" }
 solana-faucet = { path = "../faucet", version = "1.2.30" }
