+++ cli/src/lib.rs
@@ -18,16 +18,12 @@ macro_rules! pubkey {
     };
 }
 
-#[macro_use]
 extern crate serde_derive;
 
 pub mod checks;
 pub mod cli;
-pub mod cli_output;
 pub mod cluster_query;
-pub mod display;
 pub mod nonce;
-pub mod offline;
 pub mod spend_utils;
 pub mod stake;
 pub mod test_utils;
