+++ cli/tests/nonce.rs
@@ -1,15 +1,14 @@
 use solana_cli::test_utils::check_balance;
 use solana_cli::{
     cli::{process_command, request_and_confirm_airdrop, CliCommand, CliConfig},
-    cli_output::OutputFormat,
-    nonce,
-    offline::{
-        blockhash_query::{self, BlockhashQuery},
-        parse_sign_only_reply_string,
-    },
     spend_utils::SpendAmount,
 };
-use solana_client::rpc_client::RpcClient;
+use solana_cli_output::{parse_sign_only_reply_string, OutputFormat};
+use solana_client::{
+    blockhash_query::{self, BlockhashQuery},
+    nonce_utils,
+    rpc_client::RpcClient,
+};
 use solana_core::contact_info::ContactInfo;
 use solana_core::validator::{TestValidator, TestValidatorOptions};
 use solana_faucet::faucet::run_local_faucet;
@@ -299,8 +298,8 @@ fn test_create_account_with_seed() {
     check_balance(0, &rpc_client, &to_address);
 
     // Fetch nonce hash
-    let nonce_hash = nonce::get_account(&rpc_client, &nonce_address)
-        .and_then(|ref a| nonce::data_from_account(a))
+    let nonce_hash = nonce_utils::get_account(&rpc_client, &nonce_address)
+        .and_then(|ref a| nonce_utils::data_from_account(a))
         .unwrap()
         .blockhash;
