+++ cli/tests/stake.rs
@@ -1,15 +1,14 @@
 use solana_cli::test_utils::check_balance;
 use solana_cli::{
     cli::{process_command, request_and_confirm_airdrop, CliCommand, CliConfig},
-    cli_output::OutputFormat,
-    nonce,
-    offline::{
-        blockhash_query::{self, BlockhashQuery},
-        parse_sign_only_reply_string,
-    },
     spend_utils::SpendAmount,
 };
-use solana_client::rpc_client::RpcClient;
+use solana_cli_output::{parse_sign_only_reply_string, OutputFormat};
+use solana_client::{
+    blockhash_query::{self, BlockhashQuery},
+    nonce_utils,
+    rpc_client::RpcClient,
+};
 use solana_core::validator::{TestValidator, TestValidatorOptions};
 use solana_faucet::faucet::run_local_faucet;
 use solana_sdk::{
@@ -502,8 +501,8 @@ fn test_nonced_stake_delegation_and_deactivation() {
     process_command(&config).unwrap();
 
     // Fetch nonce hash
-    let nonce_hash = nonce::get_account(&rpc_client, &nonce_account.pubkey())
-        .and_then(|ref a| nonce::data_from_account(a))
+    let nonce_hash = nonce_utils::get_account(&rpc_client, &nonce_account.pubkey())
+        .and_then(|ref a| nonce_utils::data_from_account(a))
         .unwrap()
         .blockhash;
 
@@ -526,8 +525,8 @@ fn test_nonced_stake_delegation_and_deactivation() {
     process_command(&config).unwrap();
 
     // Fetch nonce hash
-    let nonce_hash = nonce::get_account(&rpc_client, &nonce_account.pubkey())
-        .and_then(|ref a| nonce::data_from_account(a))
+    let nonce_hash = nonce_utils::get_account(&rpc_client, &nonce_account.pubkey())
+        .and_then(|ref a| nonce_utils::data_from_account(a))
         .unwrap()
         .blockhash;
 
@@ -742,8 +741,8 @@ fn test_stake_authorize() {
     process_command(&config).unwrap();
 
     // Fetch nonce hash
-    let nonce_hash = nonce::get_account(&rpc_client, &nonce_account.pubkey())
-        .and_then(|ref a| nonce::data_from_account(a))
+    let nonce_hash = nonce_utils::get_account(&rpc_client, &nonce_account.pubkey())
+        .and_then(|ref a| nonce_utils::data_from_account(a))
         .unwrap()
         .blockhash;
 
@@ -788,8 +787,8 @@ fn test_stake_authorize() {
     };
     assert_eq!(current_authority, online_authority_pubkey);
 
-    let new_nonce_hash = nonce::get_account(&rpc_client, &nonce_account.pubkey())
-        .and_then(|ref a| nonce::data_from_account(a))
+    let new_nonce_hash = nonce_utils::get_account(&rpc_client, &nonce_account.pubkey())
+        .and_then(|ref a| nonce_utils::data_from_account(a))
         .unwrap()
         .blockhash;
     assert_ne!(nonce_hash, new_nonce_hash);
@@ -1024,8 +1023,8 @@ fn test_stake_split() {
     check_balance(minimum_nonce_balance, &rpc_client, &nonce_account.pubkey());
 
     // Fetch nonce hash
-    let nonce_hash = nonce::get_account(&rpc_client, &nonce_account.pubkey())
-        .and_then(|ref a| nonce::data_from_account(a))
+    let nonce_hash = nonce_utils::get_account(&rpc_client, &nonce_account.pubkey())
+        .and_then(|ref a| nonce_utils::data_from_account(a))
         .unwrap()
         .blockhash;
 
@@ -1281,8 +1280,8 @@ fn test_stake_set_lockup() {
     check_balance(minimum_nonce_balance, &rpc_client, &nonce_account_pubkey);
 
     // Fetch nonce hash
-    let nonce_hash = nonce::get_account(&rpc_client, &nonce_account.pubkey())
-        .and_then(|ref a| nonce::data_from_account(a))
+    let nonce_hash = nonce_utils::get_account(&rpc_client, &nonce_account.pubkey())
+        .and_then(|ref a| nonce_utils::data_from_account(a))
         .unwrap()
         .blockhash;
 
@@ -1400,8 +1399,8 @@ fn test_offline_nonced_create_stake_account_and_withdraw() {
     process_command(&config).unwrap();
 
     // Fetch nonce hash
-    let nonce_hash = nonce::get_account(&rpc_client, &nonce_account.pubkey())
-        .and_then(|ref a| nonce::data_from_account(a))
+    let nonce_hash = nonce_utils::get_account(&rpc_client, &nonce_account.pubkey())
+        .and_then(|ref a| nonce_utils::data_from_account(a))
         .unwrap()
         .blockhash;
 
@@ -1451,8 +1450,8 @@ fn test_offline_nonced_create_stake_account_and_withdraw() {
     check_balance(50_000, &rpc_client, &stake_pubkey);
 
     // Fetch nonce hash
-    let nonce_hash = nonce::get_account(&rpc_client, &nonce_account.pubkey())
-        .and_then(|ref a| nonce::data_from_account(a))
+    let nonce_hash = nonce_utils::get_account(&rpc_client, &nonce_account.pubkey())
+        .and_then(|ref a| nonce_utils::data_from_account(a))
         .unwrap()
         .blockhash;
 
@@ -1495,8 +1494,8 @@ fn test_offline_nonced_create_stake_account_and_withdraw() {
     check_balance(42, &rpc_client, &recipient_pubkey);
 
     // Fetch nonce hash
-    let nonce_hash = nonce::get_account(&rpc_client, &nonce_account.pubkey())
-        .and_then(|ref a| nonce::data_from_account(a))
+    let nonce_hash = nonce_utils::get_account(&rpc_client, &nonce_account.pubkey())
+        .and_then(|ref a| nonce_utils::data_from_account(a))
         .unwrap()
         .blockhash;
