+++ client/Cargo.toml
@@ -11,6 +11,7 @@ edition = "2018"
 [dependencies]
 bincode = "1.3.1"
 bs58 = "0.3.1"
+clap = "2.33.0"
 indicatif = "0.14.0"
 jsonrpc-core = "14.1.0"
 log = "0.4.8"
@@ -20,6 +21,7 @@ serde = "1.0.110"
 serde_derive = "1.0.103"
 serde_json = "1.0.53"
 solana-account-decoder = { path = "../account-decoder", version = "1.2.30" }
+solana-clap-utils = { path = "../clap-utils", version = "1.2.30" }
 solana-net-utils = { path = "../net-utils", version = "1.2.30" }
 solana-sdk = { path = "../sdk", version = "1.2.30" }
 solana-transaction-status = { path = "../transaction-status", version = "1.2.30" }
