+++ client/src/lib.rs
@@ -1,9 +1,11 @@
 #[macro_use]
 extern crate serde_derive;
 
+pub mod blockhash_query;
 pub mod client_error;
 pub mod http_sender;
 pub mod mock_sender;
+pub mod nonce_utils;
 pub mod perf_utils;
 pub mod pubsub_client;
 pub mod rpc_client;
