+++ stake-o-matic/Cargo.toml
@@ -15,6 +15,7 @@ serde_yaml = "0.8.12"
 solana-clap-utils = { path = "../clap-utils", version = "1.2.30" }
 solana-client = { path = "../client", version = "1.2.30" }
 solana-cli-config = { path = "../cli-config", version = "1.2.30" }
+solana-cli-output = { path = "../cli-output", version = "1.2.30" }
 solana-logger = { path = "../logger", version = "1.2.30" }
 solana-metrics = { path = "../metrics", version = "1.2.30" }
 solana-notifier = { path = "../notifier", version = "1.2.30" }
